Add files via upload
This code is for adding data to the database. 

Here, I fetched some details from database at first, then matched it with the new detail checking if it already exist in the database. We have to make sure that some of our details are unique (such as I.D. number etc.).

The user can also upload their passport photo